Distribute direct charm dependencies
Do not rely on system installed Python packages but distribute the
direct charm dependencies as part of the charms wheelhouse.

As the span of distributions we need to support with reactive
charms widens we will run into compability problems with the
current model.

For further reference see juju/charm-helpers#341 and
juju-solutions/layer-basic#135

Change-Id: I23c08f50f1d193a91a02e3f654202a5269ba9590