Pin tox to < 4.0.0, more pinning, add bindep.txt
Tox 4.0.0 was recently released and it has several breaking changes.
We pin to < 4.0.0 here. We are planning to move forward only on the
master charm branches.

Tox is also pinned to < 4.0.0 for stable branches in upstream
openstack-zuul-jobs as well as in zosci-config. However, the
requires= section in the charm's tox.ini file ends up installing
tox>4, wiping out the zuul-pinned tox<4 that was already installed
installed. This patch fixes that.

Update charm-tools pin to allow it to pick up future 2.x releases.

Pin PyYAML to version that doesn't use Cython.

Add bindep.txt to ensure binary dependencies are installed.

Related-Bug: #2002788
Change-Id: I3318264069b1f1bb605e054694711e8d9fe68920