Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Depends-On: I5c0b66cdbbef9ea8fcac83039bf2d1f9ccd4f804
Depends-On: https://review.opendev.org/c/openstack/charm-pacemaker-remote/+/871372
Change-Id: Ie9c07b47cd942886525c24b76f131d10c6233bfb